Fix module 05

+++ 05/ex00/Bureaucrat.cpp
@@ -15,7 +15,7 @@ Bureaucrat::Bureaucrat(): _name("Default") {
 	this->_grade = 150;
 }
 
-Bureaucrat::Bureaucrat(std::string name, uint8_t grade): _name(name) {
+Bureaucrat::Bureaucrat(std::string name, unsigned int grade): _name(name) {
 	if (grade > 150)
 		throw GradeTooLowException();
 	else if (grade < 1)
@@ -41,7 +41,7 @@ std::string Bureaucrat::getName() const {
 	return (this->_name);
 }
 
-uint8_t Bureaucrat::getGrade() const {
+unsigned int Bureaucrat::getGrade() const {
 	return (this->_grade);
 }
 
@@ -60,6 +60,6 @@ void Bureaucrat::decreaseGrade() {
 }
 
 std::ostream	&operator<<(std::ostream &os, Bureaucrat const &B) {
-	os << B.getName() << ", bureaucrat grade " << (unsigned int) B.getGrade();
+	os << B.getName() << ", bureaucrat grade " << B.getGrade();
 	return (os);
 }
