+++ 05/ex00/Bureaucrat.hpp
@@ -20,7 +20,7 @@ class Bureaucrat {
 
 public:
 	Bureaucrat();
-	Bureaucrat(std::string name, uint8_t grade);
+	Bureaucrat(std::string name, unsigned int grade);
 	Bureaucrat(Bureaucrat const &other);
 	~Bureaucrat();
 
@@ -39,14 +39,14 @@ public:
 		}
 	};
 
-	std::string	getName() const;
-	uint8_t		getGrade() const;
-	void		incrementGrade();
-	void		decreaseGrade();
+	std::string		getName() const;
+	unsigned int	getGrade() const;
+	void			incrementGrade();
+	void			decreaseGrade();
 
 private:
 	std::string	const	_name;
-	uint8_t				_grade;
+	unsigned			_grade;
 
 };
