+++ 05/ex01/Form.hpp
@@ -22,7 +22,7 @@ class Form {
 
 public:
 	Form();
-	Form(std::string name, uint8_t gradeToSign, uint8_t gradeToExecute);
+	Form(std::string name, unsigned int gradeToSign, unsigned int gradeToExecute);
 	Form(Form const &other);
 	~Form();
 
@@ -38,10 +38,10 @@ public:
 	int				getGradeToExecute() const;
 
 private:
-	std::string const	_name;
-	bool				_isSigned;
-	uint8_t const		_gradeToSign;
-	uint8_t const		_gradeToExecute;
+	std::string const		_name;
+	bool					_isSigned;
+	unsigned int const		_gradeToSign;
+	unsigned int const		_gradeToExecute;
 
 };
